+++ NBALigaSimulation/Shared/Engine/Action/Ratings/CompositeHelper.cs
@@ -23,7 +23,7 @@ namespace NBALigaSimulation.Shared.Engine
 
                     //Pace
                     player.CompositeRating.Ratings["Pace"] = 0;
-                    Dictionary<string, double> pace = new Dictionary<string, double>();
+                    Dictionary<string, decimal> pace = new Dictionary<string, decimal>();
                     pace.Add("spd", player.Ratings.LastOrDefault().Spd);
                     pace.Add("jmp", player.Ratings.LastOrDefault().Jmp);
                     pace.Add("dnk", player.Ratings.LastOrDefault().Dnk);
@@ -36,7 +36,7 @@ namespace NBALigaSimulation.Shared.Engine
 
                     //Usage
                     player.CompositeRating.Ratings["Usage"] = 0;
-                    Dictionary<string, double> usage = new Dictionary<string, double>();
+                    Dictionary<string, decimal> usage = new Dictionary<string, decimal>();
                     usage.Add("ins", player.Ratings.LastOrDefault().Ins);
                     usage.Add("dnk", player.Ratings.LastOrDefault().Dnk);
                     usage.Add("fg", player.Ratings.LastOrDefault().Fg);
@@ -46,7 +46,7 @@ namespace NBALigaSimulation.Shared.Engine
 
                     //Dribbling
                     player.CompositeRating.Ratings["Dribbling"] = 0;
-                    Dictionary<string, double> dribbling = new Dictionary<string, double>();
+                    Dictionary<string, decimal> dribbling = new Dictionary<string, decimal>();
                     dribbling.Add("drb", player.Ratings.LastOrDefault().Drb);
                     dribbling.Add("spd", player.Ratings.LastOrDefault().Spd);
                     List<string> dribblingAttributes = new List<string> { "drb", "spd" };
@@ -54,86 +54,86 @@ namespace NBALigaSimulation.Shared.Engine
 
                     //Passing
                     player.CompositeRating.Ratings["Passing"] = 0;
-                    Dictionary<string, double> passing = new Dictionary<string, double>();
+                    Dictionary<string, decimal> passing = new Dictionary<string, decimal>();
                     passing.Add("drb", player.Ratings.LastOrDefault().Drb);
                     passing.Add("pss", player.Ratings.LastOrDefault().Pss);
                     List<string> passingAttributes = new List<string> { "drb", "pss" };
-                    List<double> passingWeights = new List<double> { 0.4, 1 };
+                    List<decimal> passingWeights = new List<decimal> { 0.4m, 1 };
                     player.CompositeRating.Ratings["Passing"] = Converter.Composite(passing, passingAttributes, passingWeights);
 
                     //Turnovers
                     player.CompositeRating.Ratings["Turnovers"] = 0;
-                    Dictionary<string, double> turnovers = new Dictionary<string, double>();
+                    Dictionary<string, decimal> turnovers = new Dictionary<string, decimal>();
                     turnovers.Add("drb", player.Ratings.LastOrDefault().Drb);
                     turnovers.Add("pss", player.Ratings.LastOrDefault().Pss);
                     turnovers.Add("spd", player.Ratings.LastOrDefault().Spd);
                     turnovers.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     turnovers.Add("ins", player.Ratings.LastOrDefault().Ins);
                     List<string> turnoversAttributes = new List<string> { "drb", "pss", "spd", "hgt", "ins" };
-                    List<double> turnoversWeights = new List<double> { 1, 1, -1, 1, 1 };
+                    List<decimal> turnoversWeights = new List<decimal> { 1, 1, -1, 1, 1 };
                     player.CompositeRating.Ratings["Turnovers"] = Converter.Composite(turnovers, turnoversAttributes, turnoversWeights);
 
                     //ShootingAtRim
                     player.CompositeRating.Ratings["ShootingAtRim"] = 0;
-                    Dictionary<string, double> shootingAtRim = new Dictionary<string, double>();
+                    Dictionary<string, decimal> shootingAtRim = new Dictionary<string, decimal>();
                     shootingAtRim.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     shootingAtRim.Add("spd", player.Ratings.LastOrDefault().Spd);
                     shootingAtRim.Add("jmp", player.Ratings.LastOrDefault().Jmp);
                     shootingAtRim.Add("dnk", player.Ratings.LastOrDefault().Dnk);
                     List<string> shootingAtRimAttributes = new List<string> { "hgt", "spd", "jmp", "dnk" };
-                    List<double> shootingAtRimWeights = new List<double> { 1, 0.2, 0.6, 0.4 };
+                    List<decimal> shootingAtRimWeights = new List<decimal> { 1, 0.2m, 0.6m, 0.4m };
                     player.CompositeRating.Ratings["ShootingAtRim"] = Converter.Composite(shootingAtRim, shootingAtRimAttributes, shootingAtRimWeights);
 
                     //ShootingLowPost
                     player.CompositeRating.Ratings["ShootingLowPost"] = 0;
-                    Dictionary<string, double> shootingLowPost = new Dictionary<string, double>();
+                    Dictionary<string, decimal> shootingLowPost = new Dictionary<string, decimal>();
                     shootingLowPost.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     shootingLowPost.Add("stre", player.Ratings.LastOrDefault().Str);
                     shootingLowPost.Add("spd", player.Ratings.LastOrDefault().Spd);
                     shootingLowPost.Add("ins", player.Ratings.LastOrDefault().Ins);
                     List<string> shootingLowPostAttributes = new List<string> { "hgt", "stre", "spd", "ins" };
-                    List<double> shootingLowPostWeights = new List<double> { 1, 0.6, 0.2, 1 };
+                    List<decimal> shootingLowPostWeights = new List<decimal> { 1, 0.6m, 0.2m, 1 };
                     player.CompositeRating.Ratings["ShootingLowPost"] = Converter.Composite(shootingLowPost, shootingLowPostAttributes, shootingLowPostWeights);
 
                     //ShootingMidRange
                     player.CompositeRating.Ratings["ShootingMidRange"] = 0;
-                    Dictionary<string, double> shootingMidRange = new Dictionary<string, double>();
+                    Dictionary<string, decimal> shootingMidRange = new Dictionary<string, decimal>();
                     shootingMidRange.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     shootingMidRange.Add("fg", player.Ratings.LastOrDefault().Fg);
                     List<string> shootingMidRangeAttributes = new List<string> { "hgt", "fg" };
-                    List<double> shootingMidRangeWeights = new List<double> { 0.2, 1 };
+                    List<decimal> shootingMidRangeWeights = new List<decimal> { 0.2m, 1 };
                     player.CompositeRating.Ratings["ShootingMidRange"] = Converter.Composite(shootingMidRange, shootingMidRangeAttributes, shootingMidRangeWeights);
 
                     //ShootingThreePointer
                     player.CompositeRating.Ratings["ShootingThreePointer"] = 0;
-                    Dictionary<string, double> shootingThreePointer = new Dictionary<string, double>();
+                    Dictionary<string, decimal> shootingThreePointer = new Dictionary<string, decimal>();
                     shootingThreePointer.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     shootingThreePointer.Add("tp", player.Ratings.LastOrDefault().Tp);
                     List<string> shootingThreePointerAttributes = new List<string> { "hgt", "tp" };
-                    List<double> shootingThreePointerWeights = new List<double> { 0.2, 1 };
+                    List<decimal> shootingThreePointerWeights = new List<decimal> { 0.2m, 1 };
                     player.CompositeRating.Ratings["ShootingThreePointer"] = Converter.Composite(shootingThreePointer, shootingThreePointerAttributes, shootingThreePointerWeights);
 
                     //ShootingFT
                     player.CompositeRating.Ratings["ShootingFT"] = 0;
-                    Dictionary<string, double> shootingFT = new Dictionary<string, double>();
+                    Dictionary<string, decimal> shootingFT = new Dictionary<string, decimal>();
                     shootingFT.Add("ft", player.Ratings.LastOrDefault().Ft);
                     List<string> shootingFtAttributes = new List<string> { "ft" };
                     player.CompositeRating.Ratings["ShootingFT"] = Converter.Composite(shootingFT, shootingFtAttributes);
 
                     //Rebounding
                     player.CompositeRating.Ratings["Rebounding"] = 0;
-                    Dictionary<string, double> rebounding = new Dictionary<string, double>();
+                    Dictionary<string, decimal> rebounding = new Dictionary<string, decimal>();
                     rebounding.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     rebounding.Add("stre", player.Ratings.LastOrDefault().Str);
                     rebounding.Add("jmp", player.Ratings.LastOrDefault().Jmp);
                     rebounding.Add("reb", player.Ratings.LastOrDefault().Reb);
                     List<string> reboundingAttributes = new List<string> { "hgt", "stre", "jmp", "reb" };
-                    List<double> reboundingWeights = new List<double> { 1.5, 0.1, 0.1, 0.7 };
+                    List<decimal> reboundingWeights = new List<decimal> { 1.5m, 0.1m, 0.1m, 0.7m };
                     player.CompositeRating.Ratings["Rebounding"] = Converter.Composite(rebounding, reboundingAttributes, reboundingWeights);
 
                     //Stealing
                     player.CompositeRating.Ratings["Stealing"] = 0;
-                    Dictionary<string, double> stealing = new Dictionary<string, double>();
+                    Dictionary<string, decimal> stealing = new Dictionary<string, decimal>();
                     stealing.Add("spd", player.Ratings.LastOrDefault().Spd);
                     stealing.Add("stl", player.Ratings.LastOrDefault().Stl);
                     List<string> stealingAttributes = new List<string> { "spd", "stl" };
@@ -141,27 +141,27 @@ namespace NBALigaSimulation.Shared.Engine
 
                     //Blocking
                     player.CompositeRating.Ratings["Blocking"] = 0;
-                    Dictionary<string, double> blocking = new Dictionary<string, double>();
+                    Dictionary<string, decimal> blocking = new Dictionary<string, decimal>();
                     blocking.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     blocking.Add("jmp", player.Ratings.LastOrDefault().Jmp);
                     blocking.Add("blk", player.Ratings.LastOrDefault().Blk);
                     List<string> blockingAttributes = new List<string> { "hgt", "jmp", "blk" };
-                    List<double> blockingWeights = new List<double> { 1.5, 0.5, 0.5 };
+                    List<decimal> blockingWeights = new List<decimal> { 1.5m, 0.5m, 0.5m };
                     player.CompositeRating.Ratings["Blocking"] = Converter.Composite(blocking, blockingAttributes, blockingWeights);
 
                     //Fouling
                     player.CompositeRating.Ratings["Fouling"] = 0;
-                    Dictionary<string, double> fouling = new Dictionary<string, double>();
+                    Dictionary<string, decimal> fouling = new Dictionary<string, decimal>();
                     fouling.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     fouling.Add("blk", player.Ratings.LastOrDefault().Blk);
                     fouling.Add("spd", player.Ratings.LastOrDefault().Spd);
                     List<string> foulingAttributes = new List<string> { "hgt", "blk", "spd" };
-                    List<double> foulingWeights = new List<double> { 1, 1, -1 };
+                    List<decimal> foulingWeights = new List<decimal> { 1, 1, -1 };
                     player.CompositeRating.Ratings["Fouling"] = Converter.Composite(fouling, foulingAttributes, foulingWeights);
 
                     //Defense
                     player.CompositeRating.Ratings["Defense"] = 0;
-                    Dictionary<string, double> defense = new Dictionary<string, double>();
+                    Dictionary<string, decimal> defense = new Dictionary<string, decimal>();
                     defense.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     defense.Add("stre", player.Ratings.LastOrDefault().Str);
                     defense.Add("spd", player.Ratings.LastOrDefault().Spd);
@@ -169,52 +169,52 @@ namespace NBALigaSimulation.Shared.Engine
                     defense.Add("blk", player.Ratings.LastOrDefault().Blk);
                     defense.Add("stl", player.Ratings.LastOrDefault().Stl);
                     List<string> defenseAttributes = new List<string> { "hgt", "stre", "spd", "jmp", "blk", "stl" };
-                    List<double> defenseWeights = new List<double> { 1, 1, 1, 0.5, 1, 1 };
+                    List<decimal> defenseWeights = new List<decimal> { 1, 1, 1, 0.5m, 1, 1 };
                     player.CompositeRating.Ratings["Defense"] = Converter.Composite(defense, defenseAttributes, defenseWeights);
 
                     //DefenseInterior
                     player.CompositeRating.Ratings["DefenseInterior"] = 0;
-                    Dictionary<string, double> defenseInterior = new Dictionary<string, double>();
+                    Dictionary<string, decimal> defenseInterior = new Dictionary<string, decimal>();
                     defenseInterior.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     defenseInterior.Add("stre", player.Ratings.LastOrDefault().Str);
                     defenseInterior.Add("spd", player.Ratings.LastOrDefault().Spd);
                     defenseInterior.Add("jmp", player.Ratings.LastOrDefault().Jmp);
                     defenseInterior.Add("blk", player.Ratings.LastOrDefault().Blk);
                     List<string> defenseInteriorAttributes = new List<string> { "hgt", "stre", "spd", "jmp", "blk" };
-                    List<double> defenseInteriorWeights = new List<double> { 2, 1, 0.5, 0.5, 1 };
+                    List<decimal> defenseInteriorWeights = new List<decimal> { 2, 1, 0.5m, 0.5m, 1 };
                     player.CompositeRating.Ratings["DefenseInterior"] = Converter.Composite(defenseInterior, defenseInteriorAttributes, defenseInteriorWeights);
 
                     //DefensePerimeter
                     player.CompositeRating.Ratings["DefensePerimeter"] = 0;
-                    Dictionary<string, double> defensePerimeter = new Dictionary<string, double>();
+                    Dictionary<string, decimal> defensePerimeter = new Dictionary<string, decimal>();
                     defensePerimeter.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     defensePerimeter.Add("stre", player.Ratings.LastOrDefault().Str);
                     defensePerimeter.Add("spd", player.Ratings.LastOrDefault().Spd);
                     defensePerimeter.Add("jmp", player.Ratings.LastOrDefault().Jmp);
                     defensePerimeter.Add("stl", player.Ratings.LastOrDefault().Stl);
                     List<string> defensePerimeterAttributes = new List<string> { "hgt", "stre", "spd", "jmp", "stl" };
-                    List<double> defensePerimeterWeights = new List<double> { 1, 1, 2, 0.5, 1 };
+                    List<decimal> defensePerimeterWeights = new List<decimal> { 1, 1, 2, 0.5m, 1 };
                     player.CompositeRating.Ratings["DefensePerimeter"] = Converter.Composite(defensePerimeter, defensePerimeterAttributes, defensePerimeterWeights);
 
                     //Endurance
                     player.CompositeRating.Ratings["Endurance"] = 0;
-                    Dictionary<string, double> endurance = new Dictionary<string, double>();
+                    Dictionary<string, decimal> endurance = new Dictionary<string, decimal>();
                     endurance.Add("endu", player.Ratings.LastOrDefault().End);
                     endurance.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     List<string> enduranceAttributes = new List<string> { "endu", "hgt" };
-                    List<double> enduranceWeights = new List<double> { 1, 1, -0.1 };
+                    List<decimal> enduranceWeights = new List<decimal> { 1, 1, -0.1m };
                     player.CompositeRating.Ratings["Endurance"] = Converter.Composite(endurance, enduranceAttributes, enduranceWeights);
 
 
                     //Athleticism
                     player.CompositeRating.Ratings["Athleticism"] = 0;
-                    Dictionary<string, double> athleticism = new Dictionary<string, double>();
+                    Dictionary<string, decimal> athleticism = new Dictionary<string, decimal>();
                     athleticism.Add("stre", player.Ratings.LastOrDefault().Str);
                     athleticism.Add("spd", player.Ratings.LastOrDefault().Spd);
                     athleticism.Add("jmp", player.Ratings.LastOrDefault().Jmp);
                     athleticism.Add("hgt", player.Ratings.LastOrDefault().Hgt);
                     List<string> athleticismAttributes = new List<string> { "stre", "spd", "jmp", "hgt" };
-                    List<double> athleticismWeights = new List<double> { 1, 1, 1, 0.5 };
+                    List<decimal> athleticismWeights = new List<decimal> { 1, 1, 1, 0.5m };
                     player.CompositeRating.Ratings["Endurance"] = Converter.Composite(athleticism, athleticismAttributes, athleticismWeights);
 
                 }
@@ -241,7 +241,7 @@ namespace NBALigaSimulation.Shared.Engine
 
                     int playerRosterOrder = playersOnCourt[i][j];
                     var playerRatings = teams[i].Players.Find(player => player.RosterOrder == playerRosterOrder);
-                    double ratingValue = 0;
+                    decimal ratingValue = 0;
 
                     foreach (string rating in toUpdate)
                     {
@@ -276,12 +276,12 @@ namespace NBALigaSimulation.Shared.Engine
                     }
                 }
 
-                teams[i].CompositeRating.Ratings["GameDribbling"] += 0.1 * teams[i].Synergy.Off;
-                teams[i].CompositeRating.Ratings["GamePassing"] += 0.1 * teams[i].Synergy.Off;
-                teams[i].CompositeRating.Ratings["GameRebounding"] += 0.1 * teams[i].Synergy.Reb;
-                teams[i].CompositeRating.Ratings["GameDefense"] += 0.1 * teams[i].Synergy.Def;
-                teams[i].CompositeRating.Ratings["GameDefensePerimeter"] += 0.1 * teams[i].Synergy.Def;
-                teams[i].CompositeRating.Ratings["GameBlocking"] += 0.1 * teams[i].Synergy.Def;
+                teams[i].CompositeRating.Ratings["GameDribbling"] += 0.1m * teams[i].Synergy.Off;
+                teams[i].CompositeRating.Ratings["GamePassing"] += 0.1m * teams[i].Synergy.Off;
+                teams[i].CompositeRating.Ratings["GameRebounding"] += 0.1m * teams[i].Synergy.Reb;
+                teams[i].CompositeRating.Ratings["GameDefense"] += 0.1m * teams[i].Synergy.Def;
+                teams[i].CompositeRating.Ratings["GameDefensePerimeter"] += 0.1m * teams[i].Synergy.Def;
+                teams[i].CompositeRating.Ratings["GameBlocking"] += 0.1m * teams[i].Synergy.Def;
 
             }
         }
@@ -313,9 +313,9 @@ namespace NBALigaSimulation.Shared.Engine
             }
         }
 
-        public static double GetRatingValue(string ratingName, Player playerRatings)
+        public static decimal GetRatingValue(string ratingName, Player playerRatings)
         {
-            double ratingValue = 0;
+            decimal ratingValue = 0;
 
             switch (ratingName)
             {
